+++ Service/GlobalConstants.vb
@@ -15,6 +15,11 @@
 	Public Const CimQueryTemplateRegistryValueChange As String = "SELECT * FROM RegistryValueChangeEvent WHERE HIVE='{0}' AND KeyPath='{1}' AND ValueName='{2}'"
 #End Region 'Cim Globals
 
+#Region "Messaging Globals"
+	Public Const USUnknownError As UShort = 9999
+	Public Const StrUnknownError As String = "Unknown Error"
+#End Region
+
 #Region "Virtualization Globals"
 	Public Enum VirtualMachineStates As UShort
 		Unknown = 0