+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changes to HyperVive
 
+## 2.0: April 10, 2020
+
+- Added checkpoint watching capabilities
+- Added event IDs to all events
+- Fixed a bug that caused all WOL events to report as success
+- Enhancements to CIMitar module
+
 ## 1.0.1: March 27, 2020
 
 - Fixed a potential memory leak that might occur on a service stop.